feat: add i18n support that works as well fix multi issue with the system
