--- locales/en/common.json
@@ -1,7 +1,0 @@
-{
-    "title": "next-translate library",
-    "second-page": "Second <0>page</0>",
-    "second-page-title": "Second Page Title",
-    "loading": "Loading...",
-    "layout-title": "Layout title"
-}
