--- locales/en/home.json
@@ -1,3 +1,0 @@
-{
-    "client-only": "Only client-side"
-}
