--- locales/en/menu.json
@@ -1,6 +1,0 @@
-{
-    "guest": {
-        "sign-in-as-client": "Sign in / Register As Client",
-        "sign-in-as-delivery": "Sign in / Register As Delivery"
-    }
-}
