--- locales/en/sign-in.json
@@ -1,2 +1,0 @@
-{
-}
