+++ messages/he.json
@@ -1,0 +1,6 @@
+{
+  "menuGuest": {
+    "joinNowClient": "הצתרף כבעל עסק",
+    "joinNowDelivery": "הצתרף כשליח"
+  }
+}
