+++ next.config.mjs
@@ -1,19 +1,14 @@
 import { withPlausibleProxy } from 'next-plausible'
 import withPlugins from 'next-compose-plugins'
 import withBundleAnalyzer from '@next/bundle-analyzer'
-import nextTranslate from 'next-translate-plugin'
 import { withAxiom } from 'next-axiom';
+import withNextIntl from 'next-intl/plugin';
 
 
 const bundleAnalyzer = withBundleAnalyzer()
 /** @type {import('next').NextConfig} */
 const nextConfig = {
     productionBrowserSourceMaps: process.env.ANALYZE === 'true',
-    i18n: {
-        locales: ['en'],
-        defaultLocale: 'en',
-        localeDetection: false
-    },
     experimental: {
         missingSuspenseWithCSRBailout: false,
     },
@@ -33,8 +28,6 @@ const nextConfig = {
     }
 }
 
-const plugins = [nextTranslate({
-    reactStrictMode: true
-    }), withPlausibleProxy, withAxiom]
+const plugins = [withNextIntl, withPlausibleProxy, withAxiom]
 process.env.ANALYZE === 'true' && plugins.push(bundleAnalyzer)
 export default withPlugins(plugins, nextConfig)
