+++ package.json
@@ -79,9 +79,9 @@
     "ms": "^2.1.3",
     "next": "14.2.3",
     "next-axiom": "^1.2.0",
+    "next-intl": "^3.15.3",
     "next-joi": "^2.2.1",
     "next-plausible": "^3.12.0",
-    "next-translate": "^2.6.2",
     "postgres": "^3.4.4",
     "prop-types": "^15.8.1",
     "react": "^18",
@@ -125,7 +125,6 @@
     "eslint-plugin-storybook": "^0.8.0",
     "husky": "^9.0.11",
     "next-compose-plugins": "^2.2.1",
-    "next-translate-plugin": "^2.6.2",
     "postcss": "^8",
     "prettier": "^3.3.2",
     "storybook": "^8.1.11",
