+++ src/app/[locale]/page.tsx
@@ -1,0 +1,3 @@
+export default function Homepage() {
+  return <h1>Home Page</h1>;
+}
