+++ src/app/page.tsx
@@ -1,3 +1,5 @@
+import { redirect } from "next/navigation";
+
 export default function Homepage() {
-  return <h1>Home Page</h1>;
+  redirect("/en");
 }
